requirements.txt

python-telegram-bot==13.6
absl-py==0.9.0
astor==0.8.1
astroid==2.3.3
beautifulsoup4==4.8.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
entrypoints==0.3
flake8==3.7.9
future==0.18.2
gast==0.3.3
greenlet==0.4.15
grpcio==1.27.2
h5py==2.10.0
idna==2.9
isort==4.3.21
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
lazy-object-proxy==1.4.3
Markdown==3.2.1
markovify==0.8.0
mccabe==0.6.1
msgpack==1.0.0
numpy==1.18.1
opencv-python==4.2.0.32
protobuf==3.11.3
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
pylint==2.4.4
pynvim==0.4.1
PyYAML==5.3
requests==2.23.0
scipy==1.4.1
six==1.14.0
soupsieve==2.0
tensorboard==1.14.0
tensorflow==2.5.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
tornado==6.0.4
typed-ast==1.4.1
Unidecode==1.1.1
urllib3==1.25.8
Werkzeug==1.0.0
wrapt==1.11.2
